F-4714: compute GPT header offset in 64 bits to prevent uint32_t overflow
The protective-MBR lba_first field (attacker-controlled uint32_t, no range
check) was multiplied by GPT_SECTOR_SIZE (int 512) in disk_open. C's usual
arithmetic conversions made the product a 32-bit unsigned, which wraps for
gpt_lba >= 0x800000 (512 * 0x800001 -> 0x200), silently redirecting the GPT
header read back to LBA 1 instead of the out-of-range LBA named. Cast the
constant to uint64_t so the byte offset is computed in 64 bits. Add a unit
test that points lba_first at an overflowing LBA and confirms disk_open now
rejects it.

+++ src/disk.c
@@ -149,7 +149,8 @@ int disk_open(int drv)
         wolfBoot_printf("Found GPT PTE at sector %u\r\n", gpt_lba);
 
         /* Read GPT header */
-        r = disk_read(drv, GPT_SECTOR_SIZE * gpt_lba, GPT_SECTOR_SIZE, sector);
+        r = disk_read(drv, (uint64_t)GPT_SECTOR_SIZE * gpt_lba, GPT_SECTOR_SIZE,
+                sector);
         if (r < 0) {
             wolfBoot_printf("Disk read failed\r\n");
             Drives[drv].is_open = 0;
